Add files via upload
Create a GitHub account: If you don't have one already, sign up for a GitHub account.

Create a new repository: Once logged in, click on the "+" icon in the top-right corner of the GitHub homepage and select "New repository." Fill in the repository name, description, and other details as needed.

Initialize the repository: You have the option to initialize the repository with a README file, a .gitignore file, and a license. This can be done during the repository creation process or later.

Clone the repository: To work with the repository on your local machine, you'll need to clone it. You can do this by copying the repository URL and using the git clone command in your terminal.

Make changes: After cloning the repository, make changes to the files in your …